requirements.txt
google-generativeai>=0.3.0
flask>=2.3.0
werkzeug>=2.3.0
pillow>=9.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0